Refactor: nicer data structure for handshakes...
...
Before, handshakes were represented in the state atom as a single key with a
number - 0, 1, 2 or 3 - assoc'ed to it. This was not very nice, data-wise.
Instead, I've opted to have three different keys (per color):
:handshake-1, :handshake-2 and :handshake-3, all of them with booleans
assoc'ed to them, as with the number cards. Since I'm still persisting the
concept of up/down buttons for the handshake cards instead of the single-button
on/off toggle that I use for the number cards, more complicated functions
are needed to update the handshake's state. However, it is also very imperative
and 'simple', in the good sense, I hope - not too hard to follow. It should
probably be refactored later to be even more clear, but I'm pretty happy
with the current result.