Make :current-color implementation more transparent...
... by, instead of deriving the val from cycled colors, having :current-color
be a val in the state. This makes a `db` function redundant, so I removed it.
I also decided that it killed my need for the :last-interaction val, since
I just want the indicator to be shown for whichever color was toggled last
(via mouse or keyboard) or shifted to via the keyboard.